+++ projects/egeo/src/lib/st-fullscreen-layout/st-fullscreen-layout.ts
@@ -27,7 +27,7 @@ import { StWindowRefService } from '../utils/window-service';
  * <st-fullscreen-layout [fullWidth]="true">
  *    <p class="st-fullscreen-layout-title">Edition</p>
  *    <div class="st-fullscreen-layout-buttons">
- *       <button class="button button-secondary-line button-separator" (click)="onClickHideButton()">Cancel</button>
+ *       <button class="button button-secondary button-separator" (click)="onClickHideButton()">Cancel</button>
  *       <button class="button button-primary" (click)="onClickHideButton()">Save</button>
  *    </div>
  *    <div class="st-fullscreen-layout-content">